133 jr (#144)
* f reduce compile errors

Reduced the compile errors to a smaller set:

* `Timeout` now requires a parameter. Not really sure what that should
be. The sample code makes it look like this should be built with a
`oneshot::channel`

* the `transition!` macro may no longer work. It appears to want the
`Async` class, which no longer exists. Need to see if there's a
replacement for that.

*  `server/mod.rs` uses what looks like an older calling mechanism that
uses a `core.handle` That may no longer be required, or it may have been
converted to a channel.

* f r's

* f fix transition! macros

Still need to work on traits

* f JR's broken build for 133

* f <expletive>