Setup validation framework for encryption headers
Renamed WebPushHeaders to NotificationHeaders.

The encryption headers are only validated if there is a payload, so we
check if the payload == `Payload::None`. I've done some basic tests to
make sure this is true iff there is no payload, but to double check
I've asked in the actix-web Gitter if we can rely on this:
https://gitter.im/actix/actix-web?at=5ee295ed5dcbb760b6e40660