feat: refactor database under autocommon
WIP autoconnect needs work.

This one is not pretty.

In order to bring the database under auto_common, I have to reconcile
futures between autoconnect(0.1) and autoendpoint(0.3). This breaks the
future_state_machine as well as a number of ther dependent libraries.

This does NOT include additional work for adding in BigTable
integration, but that's going to be about the only thing missing.