feat: support dual mode data storage
Allow for dual mode, where data is written to the primary, but the
secondary is read in order to 'back fill'. For simplicity, Dual presumes
that the primary data store is Bigtable and the secondary is DynamoDb.

This presumes that the `db_settings` contains serialized settings for
both data stores.

Closes: SYNC-3452