f add flag to write to secondary (default true)
This flag makes writing to the secondary an active feature. This is
required due to the potential for clients to be arbitrarily assigned to
the dual mode storage or the older single mode storage. This flag is set
to `true` as the default since we want both data stores active until we
have sufficient cross coverage.