feat: Add integration test for mobile HTTP calls
This adds a "stub" router, specifically for integration testing. (May
put this behind a feature flag later.) Depending on the "client" type
the stub router will either succeed or fail.

This also adds some tests to ensure that mobile devices get a proper
endpoint depending on presense of a `key` value.

Closes SYNC-4041