feat: support dual mode data storage (#369)
* feat: support dual mode data storage

Allow for dual mode, where data is written to the primary, but the
secondary is read in order to 'back fill'. For simplicity, Dual presumes
that the primary data store is Bigtable and the secondary is DynamoDb.

This presumes that the `db_settings` contains serialized settings for
both data stores.

* add flag to write to secondary (default true)

This flag makes writing to the secondary an active feature. This is
required due to the potential for clients to be arbitrarily assigned to
the dual mode storage or the older single mode storage. This flag is set
to `true` as the default since we want both data stores active until we
have sufficient cross coverage.


Closes: SYNC-3452