bug: filter by timestamp (#548)
This includes a filter by cell timestamp. This also includes a lot of
inline comments pointing out potential hazards and pitfalls that I
experienced working on this.

NOTE: There is a (hopefully remote) possibility that a record could be
returned with a specific cell missing if that cell's timestamp has
expired. Timestamps are per cell, not per row.

Closes: SYNC-4060