feat: Add integration test for `dual` mode (#582)
This adds an integration test for "Dual" mode. Note that this uses a
database settings file `dual_test.json` to contain the doubly escaped
values. The values in this file should match those used in
`test_integration_all_rust.py` for data store configuration and set-up.

This also adds a bit more `debug` messages to indicate successful
operations. This was useful while determining a few problems with the
configuration, but are not strongly required. I can remove them if
requested.

Closes: https://mozilla-hub.atlassian.net/browse/SYNC-4111

---------

Co-authored-by: Philip Jenvey <pjenvey@underboss.org>