feat: introduce end_median for dual. (#663)
`end_median` is an optional database config argument for `dual`. Like
`median` it will filter UAIDs for migration by the last octet. (Note,
this will still allow UAIDs that match against the `median`)

This is an attempt to reduce hot tablet entries during migration phases.
It is recommended that `end_median` be slowly ramped up instead of
`median`.

Closes SYNC-4198