bug: Check to see if the user still exists before registering channels
there are reports of users registering a channel and the channel
immediately returning a 410. This does a simple check to see if the user
record still exists before registering. A user record may have been
removed for some reason while the user connection still persists and
this should catch for that.

Closes: SYNC-4294