feat: [SYNC-4365]: Report test results and coverage on CI (#744)
This PR splits the current `test` job, which runs both rust and
integration-tests, to 2 individual jobs. I have also added a job just
for the `cargo format` step. The 2 new jobs use a new cache which will
need to be generated on the first run so the builds will seem long but
should shorten down significantly once the cache is built.

The `Rust Test` job now reports code coverage. The file is quite large
but it is JSON and is uploaded as an artifact. This is done via
[cargo-llvm-cov](https://github.com/taiki-e/cargo-llvm-cov). The
Integration test job also reports its test results but it is stored
using `store_test_results`. I can also add this as an artifact if
needed.

This should get us the test metrics we want and also shorten our overall
test time on CI a bit.