feat: use GHA to build and push images to GAR (#750)
This creates a GitHub Actions workflow that builds and pushes Docker
images to Google Artifact Registry (GAR). This can either be triggered
in either of the following ways:

- Automatically through successful Circle CI tests when a new tag is
created
- Manually as a workflow_dispatch action

This PR does not break or change the existing CircleCI workflow that
builds Docker images and pushes them to Docker Hub. Once we know that
this GHA workflow is behaving as expected, we can remove those steps
from the CircleCI workflow and rename that to something else (ie.
"tests" instead of "build-test-deploy").

Unfortunately, it seems GitHub won't let me test the new workflow until
it's present in `master`. Once it's there, I can test/refine more as
needed from a branch.

This replaces https://github.com/mozilla-services/autopush-rs/pull/746.