feat: Normalize and explain TTLs maximums (#742)
This normalizes to a max channel TTL of 30 days and a max router TTL of
2 * max channel TTL.

See doc comment in autopush-common for details.

This also adds two convenience functions to convert SystemTime and UTC
millisecond offsets into formatted DateTime strings. These were useful
to identify the bug and may also be useful for various logging
statements.

There was also a bug in one of the unit tests, where we were taking the
cell timestamp for `connected_at` instead of the cell value. How this
did not fail before is a mystery.

---------

Co-authored-by: Philip Jenvey <pjenvey@underboss.org>