bug: compare tracking keys in raw form. (#863)
It is possible for the tracking key to be specified in Standard format
vs. URL safe. In addition, there are other formatting concerns when
comparing as encoded strings. To be more accurate, we should compare the
raw public key string.

NOTE: I'm going to fix the config file for the servers. This patch is
lower priority and is only to catch future misconfiguration problems.