f Store reliability state in Redis
There was a lot of learning that's gone on here.

First off, Redis 8 will be super awesome when it's available, it's not.
That said, there were several problems around how the reliability system
was using Redis to store state. This caused some transactions to fail,
as well as some states to fail to be properly recorded. There are
a number of `trace!()` messages added, including a few commented out
which may be useful for future diagnostics.

The load test was also improved, since I have learned more about how
the locust test works, as well as some of the constraints imposed by
our code management system.

Closes PUSH-546