feat: Document local development and debug for autopush
These files address how to "locally" develop and debug Autopush.
It's still a work in progress, and changes are GREATLY welcomed,
but this should make things a bit easier.

This also includes guidance for how to build if you are running a system
that has an incompatible environment (e.g. based on Debian Trixie)

Closes PUSH-583