feat: Document local development and debug for autopush (#1034)
These files address how to "locally" develop and debug Autopush.
It's still a work in progress, and changes are GREATLY welcomed,
but this should make things a bit easier.

This also includes guidance for how to build if you are running a system
that has an incompatible environment (e.g. based on Debian Trixie)

Closes [PUSH-583](https://mozilla-hub.atlassian.net/browse/PUSH-583)

[PUSH-583]:
https://mozilla-hub.atlassian.net/browse/PUSH-583?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ